Remove GetRandomText and Updater methods
This commit removes the GetRandomText and Updater methods. The GetRandomText method was responsible for retrieving a random text from a URL. The Updater method was handling automatic application updates. However, the automatic update functionality is not currently supported. Notably, an additional notification has been added in the ModernMainWindow class informing the user about the lack of update and multimedia functionalities and suggesting visiting the official Discord server for updates. Modifications to project asset paths were also made to reflect the changes.