Update MainWindow layout and add new Dashboard
This commit replaces the FluentWindow in MainWindow with a standard Window and modifies the configuration. A new Dashboard xaml and cs file have been added with FluentWindow configuration. It also cleans up and updates resources in the .csproj file, and modifies the solution file by removing redundant information. The Timer logic for loading the Dashboard after MainWindow is implemented in MainWindow.xaml.cs.