Update user interface and enhance functionality in various components
Several updates are made to the interface, including changes to the EasterEgg.xaml, Dashboard.xaml, and About.xaml. These include aesthetic enhancements, such as new color gradient and font, as well as added functionality in the About screen, with a dynamically changing list of cards, and navigational updates in the Dashboard. These changes aim to improve user experience and add to the overall functionality of the application.