Improve extraction functionality, add Clear History button
Enhancements made to the extraction functionality include adding the ability to disable and change the appearance of the Extraction button during the extraction process. This increases user clarity on the status of extraction. The commit also introduces a 'Clear History' button within the History tab, empowering users with the ability to clear their extraction history. Finally, drag and drop support for Unity packages was integrated, which allows users to add these packages easily to the extraction queue.