Added extraction management feature to Extraction.xaml
An update is made to the EasyExtract Extraction.xaml file. This update serves to improve the user interface by adding a new tab named "ManageExtractedTab" that provides a detailed view of the extracted UnityPackage files. The "ManageExtractedTab" is set as the default selected tab and it provides features to allow the user to manage extracted shader files from unitypackage effectively.