Changed Extraction Helper and Handler from static to instantiable classes
The Extraction Helper and Handler in the EasyExtract project were previously implemented as static classes. This commit changes these classes to be instantiable, in order to favor object-oriented principles and allow greater flexibility for potential future expansions of the codebase. Furthermore, usage of these classes in different methods has been revised to reflect this change.